backend/requirements.txt
# Django & Ecosystem
Django
django-allauth
django-cors-headers
djangorestframework
djangorestframework_simplejwt

# Database & related
psycopg
psycopg2-binary
sqlparse
asgiref

# Asynchronous
channels
channels-redis
daphne
redis

# Utilities
python-dotenv
requests
urllib3
certifi
charset-normalizer
idna
typing_extensions

# Image Processing
pillow


drf-yasg